requirements.txt
APScheduler==3.6.3
pyaes==1.6.1
pyasn1==0.4.8
pymongo==3.10.1
pytz==2020.1
rsa==4.6
short-url==1.2.2
six==1.15.0
telegram-garnet==0.3.0
Telethon==1.14.0
tzlocal==2.1
